refactor: 拆分命令注册入口
